move: 도메인에 따른 response 패키지 분리
