+++ src/main/java/com/jangburich/domain/point/domain/repository/PointTransactionRepository.java
@@ -2,7 +2,7 @@ package com.jangburich.domain.point.domain.repository;
 
 import com.jangburich.domain.point.domain.PointTransaction;
 import com.jangburich.domain.store.domain.Store;
-import com.jangburich.domain.store.dto.response.StoreChargeHistoryResponse;
+import com.jangburich.domain.store.dto.response.store.StoreChargeHistoryResponse;
 import com.jangburich.domain.user.domain.User;
 import java.util.List;
