+++ src/main/java/com/jangburich/domain/store/controller/StoreController.java
@@ -25,15 +25,15 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.jangburich.domain.store.domain.Category;
 import com.jangburich.domain.store.dto.request.StoreCreateRequest;
-import com.jangburich.domain.store.dto.response.StoreGetResponse;
+import com.jangburich.domain.store.dto.response.store.StoreGetResponse;
 import com.jangburich.domain.store.dto.request.StoreUpdateRequest;
-import com.jangburich.domain.store.dto.response.OrderDetailResponse;
-import com.jangburich.domain.store.dto.response.OrderGetResponse;
-import com.jangburich.domain.store.dto.response.OrderTodayResponse;
-import com.jangburich.domain.store.dto.response.PaymentGroupDetailResponse;
-import com.jangburich.domain.store.dto.response.SearchStoresResponse;
-import com.jangburich.domain.store.dto.response.StoreSearchDetailsResponse;
-import com.jangburich.domain.store.dto.response.StoreTeamResponse;
+import com.jangburich.domain.store.dto.response.order.OrderDetailResponse;
+import com.jangburich.domain.store.dto.response.order.OrderGetResponse;
+import com.jangburich.domain.store.dto.response.order.OrderTodayResponse;
+import com.jangburich.domain.store.dto.response.payment.PaymentGroupDetailResponse;
+import com.jangburich.domain.store.dto.response.store.SearchStoresResponse;
+import com.jangburich.domain.store.dto.response.store.StoreSearchDetailsResponse;
+import com.jangburich.domain.store.dto.response.store.StoreTeamResponse;
 import com.jangburich.domain.store.service.StoreService;
 import com.jangburich.global.payload.Message;
 import com.jangburich.global.payload.ResponseCustom;
