+++ src/main/java/com/jangburich/domain/store/dto/response/order/OrderTodayResponse.java
@@ -1,4 +1,4 @@
-package com.jangburich.domain.store.dto.response;
+package com.jangburich.domain.store.dto.response.order;
 
 import lombok.Builder;
 import lombok.Getter;
