+++ src/main/java/com/jangburich/domain/store/dto/response/payment/PrepaymentInfoResponse.java
@@ -1,4 +1,4 @@
-package com.jangburich.domain.store.dto.response;
+package com.jangburich.domain.store.dto.response.payment;
 
 import lombok.Builder;
 import lombok.Getter;
