+++ src/main/java/com/jangburich/domain/store/dto/response/store/StoreMenu.java
@@ -1,4 +1,4 @@
-package com.jangburich.domain.store.dto.response;
+package com.jangburich.domain.store.dto.response.store;
 
 import com.querydsl.core.annotations.QueryProjection;
