+++ src/main/java/com/jangburich/domain/store/dto/response/store/StoreTeamResponse.java
@@ -1,4 +1,4 @@
-package com.jangburich.domain.store.dto.response;
+package com.jangburich.domain.store.dto.response.store;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
