+++ src/main/java/com/jangburich/domain/store/repository/StoreQueryDslRepository.java
@@ -1,7 +1,7 @@
 package com.jangburich.domain.store.repository;
 
 import com.jangburich.domain.store.domain.Category;
-import com.jangburich.domain.store.dto.response.SearchStoresResponse;
+import com.jangburich.domain.store.dto.response.store.SearchStoresResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
