+++ src/main/java/com/jangburich/domain/store/repository/StoreQueryDslRepositoryImpl.java
@@ -2,7 +2,7 @@ package com.jangburich.domain.store.repository;
 
 import com.jangburich.domain.store.domain.Category;
 import com.jangburich.domain.store.dto.response.QSearchStoresResponse;
-import com.jangburich.domain.store.dto.response.SearchStoresResponse;
+import com.jangburich.domain.store.dto.response.store.SearchStoresResponse;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQuery;
