+++ src/main/java/com/jangburich/domain/store/repository/StoreTeamRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jangburich.domain.store.domain.Store;
 import com.jangburich.domain.store.domain.StoreTeam;
-import com.jangburich.domain.store.dto.response.StoreTeamResponseDTO;
+import com.jangburich.domain.store.dto.response.store.StoreTeamResponseDTO;
 
 public interface StoreTeamRepository extends JpaRepository<StoreTeam, Long> {
 	Optional<StoreTeam> findByStoreIdAndTeamId(Long store_id, Long team_id);
